Adding v_format validator
