+++ py-utils/src/utils/cli_framework/command.py
@@ -1,5 +1,4 @@
 # CORTX-Py-Utils: CORTX Python common library.
-# CORTX-Py-Utils: CORTX Python common library.
 # Copyright (c) 2021 Seagate Technology LLC and/or its Affiliates
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as published
@@ -77,114 +76,6 @@ class Command:
             output_type=self._options.get("format","success"))
 
 
-class CommandParser:
-    """
-    This Class Parses the Commands from the dictionary object
-    """
-
-    def __init__(self, cmd_data: Dict, permissions: Dict):
-        self.command = cmd_data
-        self.permissions = permissions
-        self._communication_obj = {}
-
-    def handle_main_parse(self, subparsers):
-        """
-        This Function Handles the Parsing of Single-Level and Multi-Level
-        Command Arguments
-        :param subparsers: argparser Object
-        :return:
-        """
-        if "sub_commands" in self.command:
-            self.handle_subparsers(subparsers, self.command,
-                                   self.command["name"])
-        elif "args" in self.command:
-            self.add_args(self.command, subparsers, self.command["name"])
-
-    def handle_subparsers(self, sub_parser, data: Dict, name, add_parser_flag=True):
-        """
-        This Function will handle multiple sub-parsing commands
-        :param sub_parser: Arg-parser Object
-        :param data: Data for parsing the commands :type: Dict
-        :param name: Name of the Command :type:Str
-        :param add_parser_flag: Parser Flag Decides whether an new
-            parser is object is required or not.:type:Bool
-        :return: None
-        """
-        if add_parser_flag:
-             sub_parser = sub_parser.add_parser(data["name"],
-                        help=data["description"])
-        parser = sub_parser.add_subparsers()
-        for each_data in data["sub_commands"]:
-            self.add_args(each_data, parser, name)
-
-    def check_permissions(self, sub_command):
-        """
-        filter sub_command if found any permissions tag
-        if no permissions tag is found it returns true
-        """
-        allowed = False
-        permission_tag =  sub_command.get(const.SUB_COMMANDS_PERMISSIONS, False)
-        if permission_tag and self.permissions.get(permission_tag, False):
-            allowed = True
-        return allowed
-
-    def handle_comm(self, each_args):
-        """
-        This method will handle the rest params and create the necessary object.
-        :param each_args:
-        :return:
-        """
-        if each_args.get("params", False):
-            each_args.pop("params")
-            self._communication_obj["params"][
-                each_args.get("dest", None) or each_args.get("flag")] = ""
-        if each_args.get("json", False):
-            each_args.pop("json")
-            self._communication_obj["json"][
-                each_args.get("dest", None) or each_args.get("flag")] = ""
-
-    def add_args(self, sub_command: Dict, parser: Any, name):
-        """
-        This Function will add the cmd_args from the Json to the structure.
-        :param sub_command: Action for which the command needs to be added
-        :type: str
-        :param parser: ArgParse Parser Object :type: Any
-        :param name: Name of the Command :type: str
-        :return: None
-        """
-        if not self.check_permissions(sub_command):
-            return None
-        sub_parser = parser.add_parser(sub_command["name"],
-                                       help=sub_command["description"])
-        # Check if the command has any arguments.
-        if "args" in sub_command:
-            self._communication_obj.update(sub_command["comm"])
-            self._communication_obj["params"] = {}
-            self._communication_obj["json"] = {}
-            for each_args in sub_command["args"]:
-                if each_args.get("type", None):
-                    if each_args.get("type_target"):
-                        module_obj = import_module(each_args.pop("type_target"))
-                        each_args["type"] = eval(f"module_obj.{each_args['type']}")
-                    else:
-                        each_args["type"] = eval(each_args["type"])
-                if each_args.get("suppress_help", False):
-                    each_args.pop("suppress_help")
-                    each_args["help"] = argparse.SUPPRESS
-                self.handle_comm(each_args)
-                flag = each_args.pop("flag")
-                sub_parser.add_argument(flag, **each_args)
-            sub_parser.set_defaults(command=Command,
-                                    action=deepcopy(name),
-                                    comm=deepcopy(self._communication_obj),
-                                    output=deepcopy(sub_command.get("output", {})),
-                                    need_confirmation=sub_command.get("need_confirmation", False),
-                                    sub_command_name=sub_command.get("name"))
-
-        # Check if the command has any Commands.
-        elif "sub_commands" in sub_command:
-            self.handle_subparsers(sub_parser, sub_command, name, False)
-
 
 class Output:
     """CLI Response Display Class"""
@@ -211,11 +102,11 @@ class Output:
             out.write(f"{output}\n")
 
     @staticmethod
-    def dump_success(output: Any, **kwargs):
+    def dump_success(output: Any, **kwargs) -> str:
         """
         Accepts String as Output and Returns the Same.
         :param output: Output String
-        :return:
+        :return: str
         """
         return str(kwargs.get("success", output))
